requirements.txt
streamlit
langchain
google-generativeai
faiss-cpu
sentence-transformers
gtts
streamlit_mic_recorder
python-dotenv